Module Sarek_ir_types

Sarek_ir_types - Pure type definitions for GPU kernel IR

This module contains only type definitions with no external dependencies. Used by spoc_framework for typed generate_source signature.

type memspace =

| Global

| Shared

| Local

Memory spaces

type elttype =

| TInt32

| TInt64

| TFloat32

| TFloat64

| TBool

| TUnit

| TRecord of string * (string * elttype) list
(*
Record type: name and field list
*)

| TVariant of string * (string * elttype list) list
(*
Variant type: name and constructor list with arg types
*)

| TArray of elttype * memspace
(*
Array type with element type and memory space
*)

| TVec of elttype
(*
Vector (GPU array parameter)
*)

Element types

type var = {

var_name : string;

var_id : int;

var_type : elttype;

var_mutable : bool;

}

Variables with type info

type const =

| CInt32 of int32

| CInt64 of int64

| CFloat32 of float

| CFloat64 of float

| CBool of bool

| CUnit

Constants

type binop =

| Add

| Sub

| Mul

| Div

| Mod

| Eq

| Ne

| Lt

| Le

| Gt

| Ge

| And

| Or

| Shl

| Shr

| BitAnd

| BitOr

| BitXor

Binary operators

type unop =

| Neg

| Not

| BitNot

Unary operators

type for_dir =

| Upto

| Downto

Loop direction

type pattern =

| PConstr of string * string list

| PWild

Match pattern

type expr =

| EConst of const

| EVar of var

| EBinop of binop * expr * expr

| EUnop of unop * expr

| EArrayRead of string * expr
(*
arridx
*)

| EArrayReadExpr of expr * expr
(*
base_expridx for complex bases
*)

| ERecordField of expr * string
(*
r.field
*)

| EIntrinsic of string list * string * expr list
(*
module path, name, args
*)

| ECast of elttype * expr

| ETuple of expr list

| EApp of expr * expr list

| ERecord of string * (string * expr) list
(*
Record construction: type name, field values
*)

| EVariant of string * string * expr list
(*
Variant construction: type name, constructor, args
*)

| EArrayLen of string
(*
Array length intrinsic
*)

| EArrayCreate of elttype * expr * memspace
(*
elem type, size, memspace
*)

| EIf of expr * expr * expr
(*
condition, then, else - value-returning if
*)

| EMatch of expr * (pattern * expr) list
(*
scrutinee, cases - value-returning match
*)

Expressions (pure, no side effects)

type lvalue =

| LVar of var

| LArrayElem of string * expr

| LArrayElemExpr of expr * expr

| LRecordField of lvalue * string

L-values (assignable locations)

type stmt =

| SAssign of lvalue * expr

| SSeq of stmt list

| SIf of expr * stmt * stmt option

| SWhile of expr * stmt

| SFor of var * expr * expr * for_dir * stmt

| SMatch of expr * (pattern * stmt) list

| SReturn of expr

| SBarrier
(*
Block-level barrier (__syncthreads)
*)

| SWarpBarrier
(*
Warp-level sync (__syncwarp)
*)

| SExpr of expr
(*
Side-effecting expression
*)

| SEmpty

| SLet of var * expr * stmt
(*
Let binding: let v = e in body
*)

| SLetMut of var * expr * stmt
(*
Mutable let: let v = ref e in body
*)

| SPragma of string list * stmt
(*
Pragma hints wrapping a statement
*)

| SMemFence
(*
Memory fence (threadfence)
*)

| SBlock of stmt
(*
Scoped block - creates a C scope for variable isolation
*)

| SNative of {

gpu : framework:string -> string;
(*
Generate GPU code for framework
*)

ocaml : ocaml_closure;
(*
Typed OCaml fallback
*)

}
(*
Inline native GPU code with OCaml fallback
*)

Statements (imperative, side effects)

and decl =

| DParam of var * array_info option

| DLocal of var * expr option

| DShared of string * elttype * expr option

Declarations

and array_info = {

arr_elttype : elttype;

arr_memspace : memspace;

}

and helper_func = {

hf_name : string;

hf_params : var list;

hf_ret_type : elttype;

hf_body : stmt;

}

Helper function (device function called from kernel)

and native_arg =

| NA_Int32 of int32

| NA_Int64 of int64

| NA_Float32 of float

| NA_Float64 of float

| NA_Vec of {

length : int;

elem_size : int;

type_name : string;

get_f32 : int -> float;

set_f32 : int -> float -> unit;

get_f64 : int -> float;

set_f64 : int -> float -> unit;

get_i32 : int -> int32;

set_i32 : int -> int32 -> unit;

get_i64 : int -> int64;

set_i64 : int -> int64 -> unit;

get_any : int -> Stdlib.Obj.t;

set_any : int -> Stdlib.Obj.t -> unit;

get_vec : unit -> Stdlib.Obj.t;

}

Native argument type for kernel execution. Typed arguments without Obj.t - used by PPX-generated native functions.

and ocaml_closure = {

run : block:(int * int * int) -> grid:(int * int * int) -> native_arg array -> unit;

}

Typed Helpers for Custom Types

These functions encapsulate the Obj operations so that PPX-generated code doesn't need to use Obj directly. The type parameter is inferred from context.

val vec_get_custom : 'a. native_arg -> int -> 'a

Get element from NA_Vec as custom type. Type is inferred from usage.

val vec_set_custom : 'a. native_arg -> int -> 'a -> unit

Set element in NA_Vec from custom type. Type is inferred from usage.

val vec_length : native_arg -> int

Get length from NA_Vec

val vec_as_vector : 'a. native_arg -> 'a

Get underlying vector. Used when passing vectors to functions/intrinsics that need the actual Vector.t type. Returns type-erased value that the caller casts to the appropriate Vector.t type.

type native_fn_t =

| NativeFn of parallel:bool -> block:(int * int * int) -> grid:(int * int * int) -> native_arg array -> unit

Native function type for V2 execution. Uses typed native_arg.

type kernel = {

kern_name : string;

kern_params : decl list;

kern_locals : decl list;

kern_body : stmt;

kern_types : (string * (string * elttype) list) list;
(*
Record type definitions: (type_name, (field_name, field_type); ...)
*)

kern_variants : (string * (string * elttype list) list) list;
(*
Variant type definitions: (type_name, (constructor_name, payload_types); ...)
*)

kern_funcs : helper_func list;
(*
Helper functions defined in kernel scope
*)

kern_native_fn : native_fn_t option;
(*
Optional pre-compiled native function for CPU execution
*)

}

Kernel representation