Module Spoc_framework.Typed_value

Primitive Storage

Efficient runtime representations for common types. Exotic types (Float128, SIMD) use PBytes.

type primitive =

| PInt32 of int32

| PInt64 of int64

| PFloat of float
(*
Used by both float32 and float64
*)

| PBool of bool

| PBytes of bytes
(*
For exotic scalars: Float128, SIMD, etc.
*)

Scalar Type Interface

Each intrinsic type (float32, float64, etc.) implements this interface. PPX generates these automatically for %sarek_intrinsic.

module type SCALAR_TYPE = sig ... end

Composite Type Interface

For records and structured types. Uses bytes for field serialization to avoid circular dependency with typed_value. PPX generates these automatically for @@sarek.type.

type field_desc = {

fd_name : string;
(*
Field name
*)

fd_type : string;
(*
Type name for lookup
*)

fd_offset : int;
(*
Byte offset in serialized form
*)

fd_size : int;
(*
Field size in bytes
*)

}

Field descriptor

module type COMPOSITE_TYPE = sig ... end

Typed Values

Existential wrappers for type-safe value storage.

type scalar_value =

| SV : (module SCALAR_TYPE with type t = 'a) * 'a -> scalar_value

Scalar value with its type module

type composite_value =

| CV : (module COMPOSITE_TYPE with type t = 'a) * 'a -> composite_value

Composite value with its type module

type typed_value =

| TV_Scalar of scalar_value

| TV_Composite of composite_value

Unified typed value

Vector Interface

Abstract interface for vectors, allowing execute_direct to access vector contents without depending on the full Vector module.

module type EXEC_VECTOR = sig ... end

Execution Arguments

Type-safe kernel arguments for execute_direct. No Obj.t - fully typed.

type exec_arg =

| EA_Int32 of int32

| EA_Int64 of int64

| EA_Float32 of float

| EA_Float64 of float

| EA_Scalar : (module SCALAR_TYPE with type t = 'a) * 'a -> exec_arg

| EA_Composite : (module COMPOSITE_TYPE with type t = 'a) * 'a -> exec_arg

| EA_Vec of (module EXEC_VECTOR)

Type Registry

Global registry for type lookup by name.

module Registry : sig ... end

Built-in Scalar Types

Core scalar types always available.

module Int32_type : SCALAR_TYPE with type t = int32

module Int64_type : SCALAR_TYPE with type t = int64

module Float32_type : SCALAR_TYPE with type t = float

module Float64_type : SCALAR_TYPE with type t = float

module Bool_type : SCALAR_TYPE with type t = bool

Convenience Functions

val exec_arg_of_typed_value : typed_value -> exec_arg

Create exec_arg from typed_value

val typed_value_of_exec_arg : exec_arg -> typed_value

Create typed_value from exec_arg (for scalars/composites only)

val primitive_of_scalar : scalar_value -> primitive

Get primitive value from scalar_value

val primitive_type_name : primitive -> string

Get type name from primitive

val typed_value_type_name : typed_value -> string

Get type name from typed_value

val type_name_of_exec_arg : exec_arg -> string

Get type name from exec_arg