Module type Typed_value.EXEC_VECTOR

val length : int

Number of elements

val type_name : string

Element type name

val get : int -> typed_value

Get element as typed_value

val set : int -> typed_value -> unit

Set element from typed_value

val device_ptr : unit -> nativeint

Get raw device pointer (for kernel binding)

val elem_size : int

Element size in bytes

val internal_get_vector_obj : unit -> Stdlib.Obj.t

INTERNAL: Get underlying vector as Obj.t. Only for use by interpreter backend which needs access to the typed Vector.t. This is marked internal to discourage general use - prefer the typed get/set interface.