spoc index

Library spoc.framework

The entry point of this library is the module: Spoc_framework.

Library spoc.ir

This library exposes the following toplevel modules:

Sarek_ir_analysis Sarek_ir_analysis - Analysis functions for GPU kernel IR

Sarek_ir_pp Sarek_ir_pp - Pretty printing for GPU kernel IR

Sarek_ir_types Sarek_ir_types - Pure type definitions for GPU kernel IR

Library spoc.registry

The entry point of this library is the module: Sarek_registry.